finish lesson #18

+++ app/Http/Controllers/IdeaController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Http\Requests\IdeaRequest;
 use App\Models\Idea;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -13,7 +15,8 @@ class IdeaController extends Controller
      */
     public function index()
     {
-        $ideas = Idea::all();
+        $ideas = Auth::user()->ideas;
+
         return view('ideas.index', ['ideas' => $ideas]);
     }
 
@@ -30,7 +33,10 @@ class IdeaController extends Controller
      */
     public function store(IdeaRequest $request)
     {
-        Idea::create(['description' => request('description'), 'state' => 'pending']);
+        Auth::user()->ideas()->create([
+            'description' => request('description'),
+            'state' => 'pending',
+        ]);
 
         return redirect('/ideas');
     }
@@ -40,6 +46,8 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+        Gate::authorize('update',$idea);
+
         return view('ideas.show', ['idea' => $idea]);
     }
 
@@ -48,6 +56,8 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea)
     {
+        Gate::authorize('update',$idea);
+
         return view('ideas.edit', ['idea' => $idea]);
     }
 
@@ -56,6 +66,8 @@ class IdeaController extends Controller
      */
     public function update(IdeaRequest $request, Idea $idea)
     {
+        Gate::authorize('update',$idea);
+
         $idea->update(request(['description', 'state']));
         return redirect("/ideas/{$idea->id}");
     }
@@ -65,6 +77,8 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
+        Gate::authorize('update',$idea);
+        
         $idea->delete();
         return redirect('/ideas');
     }
